Refactor map code
Related to #10

Refactor map/tileset code into a separate class.

* Add a new file `src/utils/MapManager.ts` with a `MapManager` class to handle map/tileset code.
* Move `tilemap`, `tileset`, and `layer` properties to the `MapManager` class.
* Add `preload` and `create` methods to the `MapManager` class to handle texture loading and tilemap initialization.
* Update `src/scenes/PlayScene.ts` to import and use the `MapManager` class.
* Remove map/tileset code from the `PlayScene` class and update `preload` and `create` methods to call `mapManager` methods.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/abrie/nl9/issues/10?shareId=XXXX-XXXX-XXXX-XXXX).
